+++ test/FirstDerivative/BasicArithmeticMulDiv.C
@@ -7,60 +7,60 @@
 
 extern "C" int printf(const char* fmt, ...);
 
-int m_1(int x) {
-  int y = 4;
+float m_1(float x) {
+  float y = 4;
   return y * y; // == 0
 }
-// CHECK: int m_1_darg0(int x) {
-// CHECK-NEXT: int _d_x = 1;
-// CHECK-NEXT: int _d_y = 0;
-// CHECK-NEXT: int y = 4;
+// CHECK: float m_1_darg0(float x) {
+// CHECK-NEXT: float _d_x = 1;
+// CHECK-NEXT: float _d_y = 0;
+// CHECK-NEXT: float y = 4;
 // CHECK-NEXT: return _d_y * y + y * _d_y;
 // CHECK-NEXT: }
 
-int m_2(int x) {
+float m_2(float x) {
   return 1 * 1; // == 0
 }
-// CHECK: int m_2_darg0(int x) {
-// CHECK-NEXT: int _d_x = 1;
+// CHECK: float m_2_darg0(float x) {
+// CHECK-NEXT: float _d_x = 1;
 // CHECK-NEXT: return 0 * 1 + 1 * 0;
 // CHECK-NEXT: }
 
-int m_3(int x) {
+float m_3(float x) {
   return x * x; // == 2 * x
 }
-// CHECK: int m_3_darg0(int x) {
-// CHECK-NEXT: int _d_x = 1;
+// CHECK: float m_3_darg0(float x) {
+// CHECK-NEXT: float _d_x = 1;
 // CHECK-NEXT: return _d_x * x + x * _d_x;
 // CHECK-NEXT: }
 
-int m_4(int x) {
-  int y = 4;
+float m_4(float x) {
+  float y = 4;
   return x * y * x * 3 * x; // == 9 * x * x * y
 }
-// CHECK: int m_4_darg0(int x) {
-// CHECK-NEXT: int _d_x = 1;
-// CHECK-NEXT: int _d_y = 0;
-// CHECK-NEXT: int y = 4;
-// CHECK-NEXT: int _t0 = x * y;
-// CHECK-NEXT: int _t1 = _t0 * x;
-// CHECK-NEXT: int _t2 = _t1 * 3;
+// CHECK: float m_4_darg0(float x) {
+// CHECK-NEXT: float _d_x = 1;
+// CHECK-NEXT: float _d_y = 0;
+// CHECK-NEXT: float y = 4;
+// CHECK-NEXT: float _t0 = x * y;
+// CHECK-NEXT: float _t1 = _t0 * x;
+// CHECK-NEXT: float _t2 = _t1 * 3;
 // CHECK-NEXT: return (((_d_x * y + x * _d_y) * x + _t0 * _d_x) * 3 + _t1 * 0) * x + _t2 * _d_x;
 // CHECK-NEXT: }
 
-double m_5(int x) {
+double m_5(float x) {
   return 3.14 * x;
 }
-// CHECK: double m_5_darg0(int x) {
-// CHECK-NEXT: int _d_x = 1;
+// CHECK: double m_5_darg0(float x) {
+// CHECK-NEXT: float _d_x = 1;
 // CHECK-NEXT: return 0. * x + 3.1400000000000001 * _d_x;
 // CHECK-NEXT: }
 
-float m_6(int x) {
+float m_6(float x) {
   return 3.f * x;
 }
-// CHECK: float m_6_darg0(int x) {
-// CHECK-NEXT: int _d_x = 1;
+// CHECK: float m_6_darg0(float x) {
+// CHECK-NEXT: float _d_x = 1;
 // CHECK-NEXT: return 0.F * x + 3.F * _d_x;
 // CHECK-NEXT: }
 
@@ -100,7 +100,6 @@ double m_10(double x, bool flag) {
 }
 // CHECK: double m_10_darg0(double x, bool flag) {
 // CHECK-NEXT:   double _d_x = 1;
-// CHECK-NEXT:   bool _d_flag = 0;
 // CHECK-NEXT:   return flag ? (((_d_x = _d_x * 2 + x * 0) , (x *= 2)) , (_d_x * x + x * _d_x)) : (((_d_x += 0) , (x += 1)) , (_d_x * x + x * _d_x));
 // CHECK-NEXT: }
 
@@ -113,52 +112,50 @@ double m_11(double x) {
 
 // CHECK: double m_11_darg0(double x) {
 // CHECK-NEXT:   double _d_x = 1;
-// CHECK-NEXT:   const size_t _d_maxN = 0;
 // CHECK-NEXT:   const size_t maxN = 53;
 // CHECK-NEXT:   bool _t0 = maxN < {{64U|64UL}};
-// CHECK-NEXT:   const size_t _d_m = _t0 ? _d_maxN : {{0U|0UL}};
 // CHECK-NEXT:   const size_t m = _t0 ? maxN : {{64U|64UL}};
-// CHECK-NEXT:   return _d_x * m + x * _d_m;
+// CHECK-NEXT:   return _d_x * m + x * 0;
 // CHECK-NEXT: }
 
-int d_1(int x) {
-  int y = 4;
+float d_1(float x) {
+  float y = 4;
   return y / y; // == 0
 }
-// CHECK: int d_1_darg0(int x) {
-// CHECK-NEXT: int _d_x = 1;
-// CHECK-NEXT: int _d_y = 0;
-// CHECK-NEXT: int y = 4;
+// CHECK: float d_1_darg0(float x) {
+// CHECK-NEXT: float _d_x = 1;
+// CHECK-NEXT: float _d_y = 0;
+// CHECK-NEXT: float y = 4;
 // CHECK-NEXT: return (_d_y * y - y * _d_y) / (y * y);
 // CHECK-NEXT: }
 
-int d_2(int x) {
+float d_2(float x) {
   return 1 / 1; // == 0
 }
-// CHECK: int d_2_darg0(int x) {
-// CHECK-NEXT: int _d_x = 1;
+// CHECK: float d_2_darg0(float x) {
+// CHECK-NEXT: float _d_x = 1;
 // CHECK-NEXT: return (0 * 1 - 1 * 0) / (1 * 1);
 // CHECK-NEXT: }
 
-int d_3(int x) {
+float d_3(float x) {
   return x / x; // == 0
 }
-// CHECK: int d_3_darg0(int x) {
-// CHECK-NEXT: int _d_x = 1;
+// CHECK: float d_3_darg0(float x) {
+// CHECK-NEXT: float _d_x = 1;
 // CHECK-NEXT: return (_d_x * x - x * _d_x) / (x * x);
 // CHECK-NEXT: }
 
-int d_4(int x) {
-  int y = 4;
+float d_4(float x) {
+  float y = 4;
   return x / y / x / 3 / x; // == -1 / 3 / x / x / y
 }
-// CHECK: int d_4_darg0(int x) {
-// CHECK-NEXT: int _d_x = 1;
-// CHECK-NEXT: int _d_y = 0;
-// CHECK-NEXT: int y = 4;
-// CHECK-NEXT: int _t0 = x / y;
-// CHECK-NEXT: int _t1 = _t0 / x;
-// CHECK-NEXT: int _t2 = _t1 / 3;
+// CHECK: float d_4_darg0(float x) {
+// CHECK-NEXT: float _d_x = 1;
+// CHECK-NEXT: float _d_y = 0;
+// CHECK-NEXT: float y = 4;
+// CHECK-NEXT: float _t0 = x / y;
+// CHECK-NEXT: float _t1 = _t0 / x;
+// CHECK-NEXT: float _t2 = _t1 / 3;
 // CHECK-NEXT: return (((((((_d_x * y - x * _d_y) / (y * y)) * x - _t0 * _d_x) / (x * x)) * 3 - _t1 * 0) / (3 * 3)) * x - _t2 * _d_x) / (x * x);
 // CHECK-NEXT: }
 
@@ -180,94 +177,94 @@ double issue25(double x, double y) {
 //   return 1.;
 // }
 
-int md_1(int x) {
-  int y = 4;
+float md_1(float x) {
+  float y = 4;
   return x * x / x * y / y * 3 / 3; // == 1
 }
-// CHECK: int md_1_darg0(int x) {
-// CHECK-NEXT: int _d_x = 1;
-// CHECK-NEXT: int _d_y = 0;
-// CHECK-NEXT: int y = 4;
-// CHECK-NEXT: int _t0 = x * x;
-// CHECK-NEXT: int _t1 = _t0 / x;
-// CHECK-NEXT: int _t2 = _t1 * y;
-// CHECK-NEXT: int _t3 = _t2 / y;
-// CHECK-NEXT: int _t4 = _t3 * 3;
+// CHECK: float md_1_darg0(float x) {
+// CHECK-NEXT: float _d_x = 1;
+// CHECK-NEXT: float _d_y = 0;
+// CHECK-NEXT: float y = 4;
+// CHECK-NEXT: float _t0 = x * x;
+// CHECK-NEXT: float _t1 = _t0 / x;
+// CHECK-NEXT: float _t2 = _t1 * y;
+// CHECK-NEXT: float _t3 = _t2 / y;
+// CHECK-NEXT: float _t4 = _t3 * 3;
 // CHECK-NEXT: return ((((((((_d_x * x + x * _d_x) * x - _t0 * _d_x) / (x * x)) * y + _t1 * _d_y) * y - _t2 * _d_y) / (y * y)) * 3 + _t3 * 0) * 3 - _t4 * 0) / (3 * 3);
 // CHECK-NEXT: }
 
 
-int m_1_darg0(int x);
-int m_2_darg0(int x);
-int m_3_darg0(int x);
-int m_4_darg0(int x);
-double m_5_darg0(int x);
-float m_6_darg0(int x);
+float m_1_darg0(float x);
+float m_2_darg0(float x);
+float m_3_darg0(float x);
+float m_4_darg0(float x);
+double m_5_darg0(float x);
+float m_6_darg0(float x);
 double m_7_darg0(double x);
 double m_8_darg0(double x);
 double m_9_darg0(double x);
 double m_10_darg0(double x, bool flag);
 double m_11_darg0(double x);
-int d_1_darg0(int x);
-int d_2_darg0(int x);
-int d_3_darg0(int x);
-int d_4_darg0(int x);
+float d_1_darg0(float x);
+float d_2_darg0(float x);
+float d_3_darg0(float x);
+float d_4_darg0(float x);
 double issue25_darg0(double x, double y);
-int md_1_darg0(int x);
+float md_1_darg0(float x);
 
 int main () {
-  int x = 4;
+  float x = 4;
   clad::differentiate(m_1, 0);
-  printf("Result is = %d\n", m_1_darg0(1)); // CHECK-EXEC: Result is = 0
+  printf("Result is = %.2f\n", m_1_darg0(1)); // CHECK-EXEC: Result is = 0.00
 
   clad::differentiate(m_2, 0);
-  printf("Result is = %d\n", m_2_darg0(1)); // CHECK-EXEC: Result is = 0
+  printf("Result is = %.2f\n", m_2_darg0(1)); // CHECK-EXEC: Result is = 0.00
 
   clad::differentiate(m_3, 0);
-  printf("Result is = %d\n", m_3_darg0(1)); // CHECK-EXEC: Result is = 2
+  printf("Result is = %.2f\n", m_3_darg0(1)); // CHECK-EXEC: Result is = 2.00
 
   clad::differentiate(m_4, 0);
-  printf("Result is = %d\n", m_4_darg0(1)); // CHECK-EXEC: Result is = 36
+  printf("Result is = %.2f\n", m_4_darg0(1)); // CHECK-EXEC: Result is = 36.00
 
   clad::differentiate(m_5, 0);
   printf("Result is = %f\n", m_5_darg0(1)); // CHECK-EXEC: Result is = 3.14
 
   clad::differentiate(m_6, 0);
-  printf("Result is = %f\n", m_6_darg0(1)); // CHECK-EXEC: Result is = 3
+  printf("Result is = %f\n", m_6_darg0(1)); // CHECK-EXEC: Result is = 3.00
 
   clad::differentiate(m_7, 0);
-  printf("Result is = %f\n", m_7_darg0(1)); // CHECK-EXEC: Result is = 4
+  printf("Result is = %f\n", m_7_darg0(1)); // CHECK-EXEC: Result is = 4.00
 
   clad::differentiate(m_8, 0);
-  printf("Result is = %f\n", m_8_darg0(1)); // CHECK-EXEC: Result is = 4
+  printf("Result is = %f\n", m_8_darg0(1)); // CHECK-EXEC: Result is = 4.00
 
   clad::differentiate(m_9, 0);
-  printf("Result is = %f\n", m_9_darg0(1)); // CHECK-EXEC: Result is = 8
+  printf("Result is = %f\n", m_9_darg0(1)); // CHECK-EXEC: Result is = 8.00
 
   clad::differentiate(m_10, 0);
-  printf("Result is = %f\n", m_10_darg0(1, true)); // CHECK-EXEC: Result is = 8
-  printf("Result is = %f\n", m_10_darg0(1, false)); // CHECK-EXEC: Result is = 4
+  printf("Result is = %f\n", m_10_darg0(1, true)); // CHECK-EXEC: Result is = 8.00
+  printf("Result is = %f\n", m_10_darg0(1, false)); // CHECK-EXEC: Result is = 4.00
 
   clad::differentiate(m_11<64>, 0);
   printf("Result is = %f\n", m_11_darg0(1)); // CHECK-EXEC: Result is = 53
 
   clad::differentiate(d_1, 0);
-  printf("Result is = %d\n", d_1_darg0(1)); // CHECK-EXEC: Result is = 0
+  printf("Result is = %.2f\n", d_1_darg0(1)); // CHECK-EXEC: Result is = 0.00
 
   clad::differentiate(d_2, 0);
-  printf("Result is = %d\n", d_2_darg0(1)); // CHECK-EXEC: Result is = 0
+  printf("Result is = %.2f\n", d_2_darg0(1)); // CHECK-EXEC: Result is = 0.00
 
   clad::differentiate(d_3, 0);
-  printf("Result is = %d\n", d_3_darg0(1)); // CHECK-EXEC: Result is = 0
+  printf("Result is = %.2f\n", d_3_darg0(1)); // CHECK-EXEC: Result is = 0.00
 
   clad::differentiate(d_4, 0);
-  printf("Result is = %d\n", d_4_darg0(1)); // CHECK-EXEC: Result is = 0
+  printf("Result is = %.2f\n", d_4_darg0(1)); // CHECK-EXEC: Result is = -0.08
 
   //clad::differentiate(issue25, 0);
   //printf("Result is = %f\n", issue25_darg0(1.4, 2.3)); Result is = 1
 
   clad::differentiate(md_1, 0);
-  printf("Result is = %d\n", md_1_darg0(1)); // CHECK-EXEC: Result is = 1
+  printf("Result is = %.2f\n", md_1_darg0(1)); // CHECK-EXEC: Result is = 1.00
 
   return 0;
 }
